chore: reorganize leaderboard components folder structure
