+++ src/components/leaderboard/infinite-leaderboard.tsx
